% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelos.R
\name{expandir_especies}
\alias{expandir_especies}
\title{Expandir list-column de especies}
\usage{
expandir_especies(df, coluna = "especie")
}
\arguments{
\item{df}{data.frame ou tibble com a coluna a expandir}

\item{coluna}{nome da coluna (padrao: \code{"especie"})}
}
\value{
tibble com colunas adicionais \code{<coluna>_nc}, \code{<coluna>_np}, \code{<coluna>_grupo}
}
\description{
Expande uma coluna de especies em tres colunas: nc, np e grupo.
}
\details{
Expandir list-column de especies
}
\examples{
\dontrun{
fichas |> expandir_especies("especie")
}
}
