% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelos.R
\name{expandir_ponto}
\alias{expandir_ponto}
\title{Expandir list-column de ponto}
\usage{
expandir_ponto(df, coluna = "ponto")
}
\arguments{
\item{df}{data.frame ou tibble com a coluna a expandir}

\item{coluna}{nome da coluna (padrao: \code{"ponto"})}
}
\value{
tibble com colunas adicionais \code{<coluna>_nome}, \code{<coluna>_id},
\code{<coluna>_latitude} e \code{<coluna>_longitude}
}
\description{
Expande uma coluna de ponto em nome, id, latitude e longitude.
}
\details{
Expandir list-column de ponto
}
\examples{
\dontrun{
fichas |> expandir_ponto("ponto")
}
}
