% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fichas.R
\name{ficha}
\alias{ficha}
\title{Obter uma ficha especifica}
\usage{
ficha(ficha_id)
}
\arguments{
\item{ficha_id}{string com o ID da ficha}
}
\value{
lista R com \code{id}, \code{id_app}, \code{modelo_nome}, \code{criador},
\code{created}, \code{updated}, \code{metadados_iniciais}, \code{dados}, \code{metadados_finais}
}
\description{
Retorna os dados completos de uma ficha, com os campos ja resolvidos
para seus nomes (em vez de IDs).
}
\details{
Obter uma ficha especifica
}
\examples{
\dontrun{
ficha("ficha789")
}
}
