% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fichaR-package.R
\docType{package}
\name{fichaR-package}
\alias{fichaR}
\alias{fichaR-package}
\title{fichaR: Cliente R para a API Ficharium Cloud}
\description{
Fornece funcoes para autenticacao e acesso a projetos, fichas,
modelos e especies da plataforma Ficharium.
}
\author{
\strong{Maintainer}: Biosensu \email{contato@biosensu.com}

}
