% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ficharium_base_url}
\alias{ficharium_base_url}
\title{URL base da API Ficharium}
\usage{
ficharium_base_url()
}
\value{
string com a URL base
}
\description{
Retorna a URL base da API, permitindo override via variavel de ambiente.
}
\details{
URL base da API Ficharium
}
\examples{
\dontrun{
ficharium_base_url()
}
}
