% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{ficharium_definir_token}
\alias{ficharium_definir_token}
\title{Definir token de autenticacao}
\usage{
ficharium_definir_token(token)
}
\arguments{
\item{token}{string com o token JWT}
}
\value{
invisivel
}
\description{
Salva o token JWT no \code{.Renviron} e define na sessao atual.
}
\details{
Definir token de autenticacao
}
\examples{
\dontrun{
ficharium_definir_token("meu_token_jwt")
}
}
