% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{ficharium_token}
\alias{ficharium_token}
\title{Obter token de autenticacao}
\usage{
ficharium_token()
}
\value{
string com o token
}
\description{
Le o token JWT armazenado na variavel de ambiente \code{FICHARIUM_TOKEN}.
}
\details{
Obter token de autenticacao
}
\examples{
\dontrun{
ficharium_token()
}
}
