% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projetos.R
\name{listar_projetos}
\alias{listar_projetos}
\title{Listar projetos do usuario}
\usage{
listar_projetos()
}
\value{
tibble com colunas \code{id}, \code{nome}, \code{descricao}, \code{fichas_na_nuvem}, \code{cargo}
}
\description{
Retorna todos os projetos aos quais o usuario autenticado tem acesso.
}
\details{
Listar projetos do usuario
}
\examples{
\dontrun{
ficharium_login("email@exemplo.com")
listar_projetos()
}
}
