% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projetos.R
\name{modelos_projeto}
\alias{modelos_projeto}
\title{Listar modelos de um projeto}
\usage{
modelos_projeto(projeto_id)
}
\arguments{
\item{projeto_id}{string com o ID do projeto}
}
\value{
tibble com colunas \code{id}, \code{nome}, \code{descricao}, \code{grupo}
}
\description{
Retorna os modelos associados a um projeto especifico.
}
\details{
Listar modelos de um projeto
}
\examples{
\dontrun{
modelos_projeto("proj123")
}
}
