% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projetos.R
\name{projeto}
\alias{projeto}
\title{Obter um projeto especifico}
\usage{
projeto(id)
}
\arguments{
\item{id}{string com o ID do projeto}
}
\value{
lista R com os campos do projeto
}
\description{
Retorna os detalhes completos de um projeto pelo seu ID.
}
\details{
Obter um projeto especifico
}
\examples{
\dontrun{
projeto("abc123")
}
}
